+++ csr_cpp/Worklist.cpp
@@ -5,17 +5,14 @@
 #include "Worklist.h"
 
 Worklist::Worklist(CSR graph, int32_t delta) : csr(graph), delta(delta), relaxCount(0) {
-    buckets = map<int32_t, set<int32_t>>();
+    buckets = map<long, set<int32_t>>();
     vector<vector<int32_t>> vals = graph.iterate();
     int32_t w = 0;
 
     for(auto it = vals.begin(); it != vals.end(); ++it){
         w = it->at(0);
-        int32_t i = floor(graph.getTent(w)/delta);
-//        cout << "w = " << w << endl;
-//        cout << "i = " << i <<endl;
-
-        if(buckets.find(i) == buckets.end())buckets[i] = set<int32_t>();
+        long i = floor(graph.getTent(w)/delta);
+        if(buckets.find(i) == buckets.end()) buckets[i] = set<int32_t>();
         buckets[i].insert(w);
     }
 
@@ -42,7 +39,7 @@ bool Worklist::hasElements() {
     return false;
 }
 
-int32_t Worklist::getIndex() {
+long Worklist::getIndex() {
     for(auto it = buckets.begin(); it != buckets.end(); ++it)
         if(it->second.size() > 0) {
             cout << "key: " << it->first << endl;
@@ -52,22 +49,22 @@ int32_t Worklist::getIndex() {
     return -1;
 }
 
-set<int32_t> Worklist::get(int32_t i) {
+set<int32_t> Worklist::get(long i) {
     return buckets[i];
 }
 
-void Worklist::put(int32_t i, set<int32_t> nodes) {
+void Worklist::put(long i, set<int32_t> nodes) {
     buckets[i] = nodes;
 }
 
-void Worklist::relax(int32_t w, int32_t d) {
+void Worklist::relax(int32_t w, long d) {
     ++relaxCount;
-    int32_t tentCost = csr.getTent(w);
+    long tentCost = csr.getTent(w);
     if(d < tentCost){
         csr.setTent(w, d);
-        int32_t i = floor(tentCost/delta);
+        long i = floor(tentCost/delta);
         if(buckets[i].find(w) != buckets[i].end()) buckets[i].erase(w);
-        int32_t newInd = floor(d/delta);
+        long newInd = floor(d/delta);
         if(buckets.find(newInd) == buckets.end())
             buckets[newInd] = set<int32_t>({w});
         else buckets[newInd].insert(w);
